Add memory fact network scan fingerprints
Part of DISCOVERY-130 implementation

+++ quipucords/api/deployments_report/model.py
@@ -107,7 +107,7 @@ class SystemFingerprint(models.Model):
     cpu_count = models.PositiveIntegerField(unique=False, blank=True, null=True)
 
     architecture = models.CharField(max_length=64, unique=False, blank=True, null=True)
-    system_memory_bytes = models.PositiveIntegerField(
+    system_memory_bytes = models.PositiveBigIntegerField(
         unique=False, blank=True, null=True
     )
